+++ crates/rts-std/src/lib.rs
@@ -22,4 +22,5 @@ pub mod http_server;
 pub mod events;
 pub mod promise_slot;
 pub mod time;
+pub mod crypto;
 pub mod globals;
